test(core): Consolidate transition-rollback test into order-process suite
Move the #4686 onTransitionEnd rollback test out of the standalone
order-process-rollback.e2e-spec.ts file and into the existing
order-process.e2e-spec.ts as a final describe block. Uses a
toggle-based failingOrderProcess so the existing 13 tests in the
file are unaffected — the toggle defaults to null and is only
enabled by the rollback test's setup.

Targets the ValidatingCustomer state for the failure injection
because that is the only target reachable from AddingItems under
the existing customOrderProcess (the 'replace' merge strategy
collapses default transitions on AddingItems).

Relates to #4686
